+++ pg-core/src/java/org/pg/enums/ConnType.java
@@ -1,0 +1,5 @@
+package org.pg.enums;
+
+public enum ConnType {
+    INET4, UNIX_SOCKET
+}