+++ pg-core/src/java/org/pg/enums/SSLValidation.java
@@ -1,0 +1,5 @@
+package org.pg.enums;
+
+public enum SSLValidation {
+    NONE, DEFAULT
+}